feat: upload to Cloudinary for Medicine and Doctor

+++ src/components/Manage_Staff/edit_staff.vue
@@ -47,7 +47,7 @@ export default {
             return publicId;
         },
         async editStaff(data) {
-            const result = await Swal.fire({
+            await Swal.fire({
                 title: "Bạn muốn cập nhật tài khoản này?",
                 showDenyButton: true,
                 showCancelButton: true,
@@ -117,8 +117,10 @@ export default {
                         }
                     }
                     catch (error) {
-                        const publicID = await this.separateLinkIntoPublicID(this.imgURLCloud)
-                        await CloudService.delete(publicID);
+                        if (data.imgURL instanceof File) {
+                            const publicID = await this.separateLinkIntoPublicID(this.imgURLCloud)
+                            await CloudService.delete(publicID);
+                        }
                         this.handleError(error);
                     } finally {
                         this.loading = false;
@@ -150,69 +152,7 @@ export default {
     },
 };
 </script>
+
 <style scoped>
-/* ---------------------------- LOADER ---------------------------- */
-.d-load {
-    min-width: 100%;
-    height: 100px;
-}
-
-.loader {
-    margin: 180px auto 0px;
-    width: 50px;
-    aspect-ratio: 1;
-    border-radius: 50%;
-    border: 8px solid #514b82;
-    animation:
-        l20-1 0.8s infinite linear alternate,
-        l20-2 1.6s infinite linear;
-}
-
-@keyframes l20-1 {
-    0% {
-        clip-path: polygon(50% 50%, 0 0, 50% 0%, 50% 0%, 50% 0%, 50% 0%, 50% 0%)
-    }
-
-    12.5% {
-        clip-path: polygon(50% 50%, 0 0, 50% 0%, 100% 0%, 100% 0%, 100% 0%, 100% 0%)
-    }
-
-    25% {
-        clip-path: polygon(50% 50%, 0 0, 50% 0%, 100% 0%, 100% 100%, 100% 100%, 100% 100%)
-    }
-
-    50% {
-        clip-path: polygon(50% 50%, 0 0, 50% 0%, 100% 0%, 100% 100%, 50% 100%, 0% 100%)
-    }
-
-    62.5% {
-        clip-path: polygon(50% 50%, 100% 0, 100% 0%, 100% 0%, 100% 100%, 50% 100%, 0% 100%)
-    }
-
-    75% {
-        clip-path: polygon(50% 50%, 100% 100%, 100% 100%, 100% 100%, 100% 100%, 50% 100%, 0% 100%)
-    }
-
-    100% {
-        clip-path: polygon(50% 50%, 50% 100%, 50% 100%, 50% 100%, 50% 100%, 50% 100%, 0% 100%)
-    }
-}
-
-@keyframes l20-2 {
-    0% {
-        transform: scaleY(1) rotate(0deg)
-    }
-
-    49.99% {
-        transform: scaleY(1) rotate(135deg)
-    }
-
-    50% {
-        transform: scaleY(-1) rotate(0deg)
-    }
-
-    100% {
-        transform: scaleY(-1) rotate(-135deg)
-    }
-}
+@import "@/assets/css/formCRUD.css";
 </style>
